+++ src/main/java/com/verdantartifice/primalmagick/common/research/ResearchEntry.java
@@ -39,6 +39,8 @@ public class ResearchEntry {
     protected ResearchEntry.Icon icon;
     protected CompoundResearchKey parentResearch;
     protected boolean hidden;
+    protected boolean finaleExempt;
+    protected List<String> finales = new ArrayList<>();
     protected List<ResearchStage> stages = new ArrayList<>();
     protected List<ResearchAddendum> addenda = new ArrayList<>();
     
@@ -47,6 +49,7 @@ public class ResearchEntry {
         this.disciplineKey = disciplineKey;
         this.nameTranslationKey = nameTranslationKey;
         this.hidden = false;
+        this.finaleExempt = false;
         this.icon = icon;
     }
     
@@ -77,8 +80,18 @@ public class ResearchEntry {
             entry.hidden = obj.getAsJsonPrimitive("hidden").getAsBoolean();
         }
         
+        if (obj.has("finaleExempt")) {
+            entry.finaleExempt = obj.getAsJsonPrimitive("finaleExempt").getAsBoolean();
+        }
+        
         if (obj.has("parents")) {
             entry.parentResearch = CompoundResearchKey.parse(obj.get("parents").getAsJsonArray());
+        }
+        
+        if (obj.has("finales")) {
+            for (JsonElement element : obj.get("finales").getAsJsonArray()) {
+                entry.finales.add(element.getAsString());
+            }
         }
         
         for (JsonElement element : obj.get("stages").getAsJsonArray()) {
@@ -102,7 +115,12 @@ public class ResearchEntry {
         ResearchEntry.Icon icon = ResearchEntry.Icon.fromNetwork(buf);
         ResearchEntry entry = create(key, discipline, name, icon);
         entry.hidden = buf.readBoolean();
+        entry.finaleExempt = buf.readBoolean();
         entry.parentResearch = CompoundResearchKey.parse(buf.readUtf());
+        int finaleCount = buf.readVarInt();
+        for (int index = 0; index < finaleCount; index++) {
+            entry.finales.add(buf.readUtf());
+        }
         int stageCount = buf.readVarInt();
         for (int index = 0; index < stageCount; index++) {
             entry.stages.add(ResearchStage.fromNetwork(buf, entry));
@@ -120,7 +138,12 @@ public class ResearchEntry {
         buf.writeUtf(entry.nameTranslationKey);
         ResearchEntry.Icon.toNetwork(buf, entry.icon);
         buf.writeBoolean(entry.hidden);
+        buf.writeBoolean(entry.finaleExempt);
         buf.writeUtf(entry.parentResearch == null ? "" : entry.parentResearch.toString());
+        buf.writeVarInt(entry.finales.size());
+        for (String discipline : entry.finales) {
+            buf.writeUtf(discipline);
+        }
         buf.writeVarInt(entry.stages.size());
         for (ResearchStage stage : entry.stages) {
             ResearchStage.toNetwork(buf, stage);
@@ -158,6 +181,36 @@ public class ResearchEntry {
     
     public boolean isHidden() {
         return this.hidden;
+    }
+    
+    /**
+     * Get whether this research entry is exempt from counting towards discipline completion for unlocking finale research.
+     * 
+     * @return whether this entry is finale exempt
+     */
+    public boolean isFinaleExempt() {
+        return this.finaleExempt;
+    }
+    
+    /**
+     * Get a list of all discipline keys for which this entry is a finale.  For this entry to be unlocked,
+     * all listed disciplines must be completed.
+     * 
+     * @return all discipline keys for which this entry is a finale
+     */
+    @Nonnull
+    public List<String> getFinaleDisciplines() {
+        return Collections.unmodifiableList(this.finales);
+    }
+    
+    /**
+     * Get whether this research entry is a finale for the given discipline key.
+     * 
+     * @param discipline the discipline to be tested
+     * @return whether this research is a finale for the given discipline key
+     */
+    public boolean isFinaleFor(String discipline) {
+        return this.finales.contains(discipline);
     }
     
     @Nonnull